requirements.txt
pandas
openpyxl
requests
PyYAML 